requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
joblib
psycopg2-binary
python-dotenv